+++ disk/tox.ini
@@ -7,9 +7,7 @@ envlist =
 
 [testenv]
 usedevelop = true
-platform = linux2|darwin
-
-[testenv:disk]
+platform = linux|darwin|win32
 deps =
     -e../datadog_checks_base[deps]
     -rrequirements-dev.txt
